Rename package as pymatcalc6
