+++ pymatcalc6/utils.py
